Create Product class and made 1 to N relation market-products

+++ backend/core/src/main/java/decidish/com/core/model/rewe/ProductPrice.java
@@ -1,8 +1,8 @@
 package decidish.com.core.model.rewe;
 
 public record ProductPrice (
-    Long currentRetailPrice,
-    Long totalRefundPrice,
+    int currentRetailPrice,
+    int totalRefundPrice,
     String grammage,
     Discount discount,
     String loyaltyBonus
